+++ src/factory/core/ports/outbound_listener.py
@@ -1,0 +1,39 @@
+"""Structural protocol for adapter-side outbound listeners.
+
+Defines the minimum surface TelegramAdapter / DiscordAdapter call on their
+cached listener. NatsOutboundListener satisfies this protocol structurally
+(no inheritance required). Keeps adapter code transport-agnostic at the
+type level.
+
+Relocated from factory.adapters.shared.outbound_listener to factory.core.ports
+so that factory.inbound can reference it without violating the
+inbound-no-adapters importlinter contract (ADR-073 / #1666).
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Protocol
+
+if TYPE_CHECKING:
+    from factory.core.messaging.message import InboundMessage
+
+
+class OutboundListener(Protocol):
+    """Surface that adapter code requires from its outbound listener.
+
+    Matches the shape of NatsOutboundListener. Not @runtime_checkable —
+    this mirrors the ChannelAdapter protocol convention in
+    core/hub/hub_protocol.py.
+    """
+
+    def cache_inbound(self, msg: InboundMessage) -> None:
+        """Store msg so outbound correlation can retrieve it by stream_id."""
+        ...
+
+    async def start(self) -> None:
+        """Start consuming outbound events from the transport."""
+        ...
+
+    async def stop(self) -> None:
+        """Stop consuming and release transport resources."""
+        ...
