+++ src/factory/typing/__init__.py
@@ -1,4 +1,5 @@
 from factory.typing.listener import TypingListener, make_typing_factory
+from factory.typing.task_manager import TypingTaskManager
 from factory.typing.types import (
     CoroFactory,
     FactoryBuilder,
@@ -12,5 +13,6 @@ __all__ = [
     "ScopeResolver",
     "TypingListener",
     "TypingManagerProtocol",
+    "TypingTaskManager",
     "make_typing_factory",
 ]
